+++ src/scip/cons_sos2.c
@@ -2249,7 +2249,7 @@ SCIP_DECL_CONSGETPERMSYMGRAPH(consGetPermsymGraphSOS2)
    /* for all pairs of variables, add a node indicating a tuple and add nodes for (aggregated) variables */
    for( i = 0; i < nconsvars - 1; ++i )
    {
-      opnodeidx = SCIPaddSymgraphOpnode(scip, graph, SYM_CONSOPTYPE_TUPLE);
+      opnodeidx = SCIPaddSymgraphOpnode(scip, graph, (int) SYM_CONSOPTYPE_TUPLE); /*lint !e641*/
       SCIP_CALL( SCIPaddSymgraphEdge(scip, graph, consnodeidx, opnodeidx, FALSE, 0.0) );
 
       for( j = i; j < i + 2; ++j )
@@ -2270,7 +2270,7 @@ SCIP_DECL_CONSGETPERMSYMGRAPH(consGetPermsymGraphSOS2)
          }
          else
          {
-            nodeidx = SCIPaddSymgraphOpnode(scip, graph, SYM_CONSOPTYPE_SUM);
+            nodeidx = SCIPaddSymgraphOpnode(scip, graph, (int) SYM_CONSOPTYPE_SUM); /*lint !e641*/
 
             SCIP_CALL( SCIPaddSymgraphEdge(scip, graph, opnodeidx, nodeidx, FALSE, 0.0) );
